fix compile error in linux
git-svn-id: https://subversive.cims.nyu.edu/exact@7918 a02fd672-c32b-43fa-a32f-4ef1cc386a6b

+++ trunk/progs/robotics/complex-general-qt/complex.h
@@ -353,7 +353,7 @@ bool loadComplexRobot(string& robot_file, vector<Triangle>& compRoboTri,
 
 
   if (verboseOption) {
-      mw_out << "Robot has "<< nv << " verts, " << compRoboTri.size() << " triangles.\n";
+      mw_out << "Robot has "<< nv << " verts, " << (int)compRoboTri.size() << " triangles.\n";
   }
   return true;
 }
